requirements.txt
anyio==4.3.0
arrow==1.3.0
asgiref==3.8.1
black==24.3.0
blessed==1.20.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
Django==4.1.3
django-picklefield==3.1

djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.0.1
redis==3.5.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
wcwidth==0.2.13
requests==2.31.0
sqlparse==0.4.4
stripe==8.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
